closes: #6904 fixed hourly usage
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/work/src@24720 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ webfilter-base/hier/usr/lib/python2.5/reports/node/untangle_base_webfilter.py
@@ -261,8 +261,8 @@ FROM (SELECT date_trunc('hour', trunc_time) AS hour,
             q = """\
 SELECT (date_part('hour', trunc_time) || ':'
         || (date_part('minute', trunc_time)::int / 10 * 10))::time AS time,
-       coalesce(sum(hits), 0)::int / (10 * %%s) AS hits,
-       coalesce(sum(wf_%s_blocks), 0)::int / (10 * %%s) AS wf_%s_blocks
+       coalesce(sum(hits), 0) / (10 * %%s) AS hits,
+       coalesce(sum(wf_%s_blocks), 0) / (10 * %%s) AS wf_%s_blocks
 FROM reports.n_http_totals
 WHERE trunc_time >= %%s AND trunc_time < %%s""" % (2 * (self.__vendor_name,))
             if host:
